tests/end2end: migrate tests to safe server code: UC07_G[1-2]
Relates to #925
