+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T03_remove_existing_link/expected_output/document.sdoc
