+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T03_remove_existing_link/test_UC07_G1_T03_remove_existing_link.py
@@ -5,7 +5,7 @@ from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
 
 
-class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
+class Test_UC07_G1_T03_RemoveLink(BaseCase):
     def test_01(self):
         test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
@@ -21,31 +21,37 @@ class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
 
             self.assert_text("Hello world!")
 
+            # Make sure that the normal (not table-based) requirement is
+            # rendered.
+            self.assert_element(
+                '//sdoc-node[@data-testid="node-requirement-normal"]',
+                by=By.XPATH,
+            )
+
             self.hover_and_click(
-                hover_selector="(//sdoc-node)[2]",
+                hover_selector="(//sdoc-node)[3]",
                 click_selector=(
-                    '(//sdoc-node)[2]//*[@data-testid="node-edit-action"]'
+                    '(//sdoc-node)[3]//*[@data-testid="node-edit-action"]'
                 ),
                 hover_by=By.XPATH,
                 click_by=By.XPATH,
             )
 
-            self.type(
-                "//*[@id='requirement[STATEMENT]']",
-                """
-- Broken RST markup
-
-  - AAA
-  ---
-    """.strip(),
-                by=By.XPATH,
+            self.click_xpath(
+                '//*[@data-testid="form-delete-'
+                'requirement[REFS_PARENT][]-field-action"]'
             )
 
+            self.scroll_to(
+                "//button[@type='submit' and text()='Save']", by=By.XPATH
+            )
             self.click_xpath('//*[@data-testid="form-submit-action"]')
 
-            self.assert_text(
-                "RST markup syntax error on line 4: "
-                "Bullet list ends without a blank line; unexpected unindent."
+            # TODO: Make sure that the link with the REQ-001 text no longer
+            # exists.
+
+            self.assert_element_not_present(
+                "//button[@type='submit' and text()='Save']", by=By.XPATH
             )
 
         assert test_setup.compare_sandbox_and_expected_output()
