+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T10_editing_requirement_refs_validate_empty/expected_output/document.sdoc
