+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T10_editing_requirement_refs_validate_empty/input/document.sdoc
@@ -1,0 +1,15 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Hello world!
+[/FREETEXT]
+
+[REQUIREMENT]
+TITLE: Requirement title
+STATEMENT: >>>
+Requirement statement.
+<<<
+RATIONALE: >>>
+Requirement rationale.
+<<<
