+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T10_editing_requirement_refs_validate_empty/test_UC07_G1_T10_editing_requirement_refs.py
@@ -5,7 +5,7 @@ from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
 
 
-class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
+class Test_UC07_G1_T10_EditRequirementRefs(BaseCase):
     def test_01(self):
         test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
@@ -20,6 +20,12 @@ class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
             self.click_xpath('//*[@data-testid="tree-file-link"]')
 
             self.assert_text("Hello world!")
+            # Make sure that the normal (not table-based) requirement is
+            # rendered.
+            self.assert_element(
+                '//sdoc-node[@data-testid="node-requirement-normal"]',
+                by=By.XPATH,
+            )
 
             self.hover_and_click(
                 hover_selector="(//sdoc-node)[2]",
@@ -30,22 +36,12 @@ class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
                 click_by=By.XPATH,
             )
 
-            self.type(
-                "//*[@id='requirement[STATEMENT]']",
-                """
-- Broken RST markup
-
-  - AAA
-  ---
-    """.strip(),
-                by=By.XPATH,
-            )
+            self.click_xpath("//*[@data-testid='form-add-parent-link-action']")
 
             self.click_xpath('//*[@data-testid="form-submit-action"]')
 
             self.assert_text(
-                "RST markup syntax error on line 4: "
-                "Bullet list ends without a blank line; unexpected unindent."
+                "Requirement parent link UID must not be empty.",
             )
 
         assert test_setup.compare_sandbox_and_expected_output()
