+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T12_added_link_when_no_uid/input/document.sdoc
@@ -1,0 +1,25 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Hello world!
+[/FREETEXT]
+
+[REQUIREMENT]
+TITLE: Requirement #1
+STATEMENT: >>>
+Requirement statement.
+<<<
+RATIONALE: >>>
+Requirement rationale.
+<<<
+
+[REQUIREMENT]
+UID: REQ-002
+TITLE: Requirement #2
+STATEMENT: >>>
+Requirement statement.
+<<<
+RATIONALE: >>>
+Requirement rationale.
+<<<
