+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T12_added_link_when_no_uid/test_UC07_G1_T12_added_link_when_no_uid.py
@@ -5,7 +5,7 @@ from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
 
 
-class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
+class Test_UC07_G1_T12_AddedLinkWhenNoUID(BaseCase):
     def test_01(self):
         test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
@@ -30,22 +30,22 @@ class Test_UC07_G2_T02_StatementMalformedRST(BaseCase):
                 click_by=By.XPATH,
             )
 
+            self.click_xpath("//*[@data-testid='form-add-parent-link-action']")
             self.type(
-                "//*[@id='requirement[STATEMENT]']",
-                """
-- Broken RST markup
-
-  - AAA
-  ---
-    """.strip(),
-                by=By.XPATH,
+                (
+                    "(//*["
+                    "@data-testid='form-requirement[REFS_PARENT][]-field'"
+                    "])[1]"
+                ),
+                "REQ-002",
             )
 
             self.click_xpath('//*[@data-testid="form-submit-action"]')
 
             self.assert_text(
-                "RST markup syntax error on line 4: "
-                "Bullet list ends without a blank line; unexpected unindent."
+                "Requirement with parent links must have an UID. "
+                "Either provide a parent UID, or "
+                "delete the parent requirement links.",
             )
 
         assert test_setup.compare_sandbox_and_expected_output()
