+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T13_renaming_uid_when_parent_links_exist/input/document.sdoc
