+++ tests/end2end/document/UC07_edit_requirement/UC07_G1_editing_requirement_links/UC07_G1_T20_two_documents_removing_link/expected_output/document2.sdoc
