+++ tests/end2end/document/UC07_edit_requirement/UC07_G2_validation/UC07_G2_T01_empty_statement/input/document.sdoc
