+++ tests/end2end/document/UC07_edit_requirement/UC07_G2_validation/UC07_G2_T01_empty_statement/test_UC07_G2_T01_empty_statement.py
@@ -1,59 +1,50 @@
-import os
-import shutil
-
 from selenium.webdriver import Keys
 from selenium.webdriver.common.by import By
 from seleniumbase import BaseCase
 
+from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
-
-path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 
 class Test_UC07_G2_T01_EmptyStatement(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        shutil.copyfile(
-            os.path.join(path_to_this_test_file_folder, "document.sdoc"),
-            os.path.join(path_to_sandbox, "document.sdoc"),
-        )
+        with SDocTestServer(
+            input_path=test_setup.path_to_sandbox
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        test_server.run()
+            self.assert_text("Document 1")
+            self.assert_text("PROJECT INDEX")
 
-        self.open(test_server.get_host_and_port())
+            self.click_xpath('//*[@data-testid="tree-file-link"]')
 
-        self.assert_text("Document 1")
-        self.assert_text("PROJECT INDEX")
+            self.assert_text("Hello world!")
 
-        self.click_xpath('//*[@data-testid="tree-file-link"]')
+            self.hover_and_click(
+                hover_selector="(//sdoc-node)[2]",
+                click_selector=(
+                    '(//sdoc-node)[2]//*[@data-testid="node-edit-action"]'
+                ),
+                hover_by=By.XPATH,
+                click_by=By.XPATH,
+            )
 
-        self.assert_text("Hello world!")
+            self.type(
+                "//*[@id='requirement[TITLE]']", "Modified title", by=By.XPATH
+            )
 
-        self.hover_and_click(
-            hover_selector="(//sdoc-node)[2]",
-            click_selector=(
-                '(//sdoc-node)[2]//*[@data-testid="node-edit-action"]'
-            ),
-            hover_by=By.XPATH,
-            click_by=By.XPATH,
-        )
+            # HACK: The only way the field is actually cleared.
+            self.type("//*[@id='requirement[STATEMENT]']", "X", by=By.XPATH)
+            requirement_statement_field = self.find_element(
+                "//*[@id='requirement[STATEMENT]']", by=By.XPATH
+            )
+            requirement_statement_field.click()
+            requirement_statement_field.send_keys(Keys.BACKSPACE)
 
-        self.type(
-            "//*[@id='requirement[TITLE]']", "Modified title", by=By.XPATH
-        )
+            self.click_xpath('//*[@data-testid="form-submit-action"]')
 
-        # HACK: The only way the field is actually cleared.
-        self.type("//*[@id='requirement[STATEMENT]']", "X", by=By.XPATH)
-        requirement_statement_field = self.find_element(
-            "//*[@id='requirement[STATEMENT]']", by=By.XPATH
-        )
-        requirement_statement_field.click()
-        requirement_statement_field.send_keys(Keys.BACKSPACE)
+            self.assert_text("Requirement statement must not be empty.")
 
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
-
-        self.assert_text("Requirement statement must not be empty.")
+        assert test_setup.compare_sandbox_and_expected_output()
