+++ tests/end2end/document/UC07_edit_requirement/UC07_G2_validation/UC07_G2_T02_statement_malformed_rst/expected_output/document.sdoc
@@ -6,11 +6,7 @@ Hello world!
 [/FREETEXT]
 
 [REQUIREMENT]
-UID: Modified UID
-TITLE: Modified title
+TITLE: Requirement title
 STATEMENT: >>>
-Modified statement.
-<<<
-RATIONALE: >>>
-Modified rationale.
+Requirement statement.
 <<<
