+++ tests/end2end/document/UC07_edit_requirement/UC07_G2_validation/UC07_G2_T02_statement_malformed_rst/input/document.sdoc
@@ -1,0 +1,12 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Hello world!
+[/FREETEXT]
+
+[REQUIREMENT]
+TITLE: Requirement title
+STATEMENT: >>>
+Requirement statement.
+<<<
